Added components for transition buffer and related stats.

+++ gateway/configuration/config.go
@@ -45,7 +45,7 @@ type GatewayConfig struct {
 	// Exporters contains the configuration for the included exporters.
 	Exporters *ExportersConfig `json:"exporters"`
 	// GatewayTransitionBufferSize tunes the size of the buffer between targets and exporters/clients.
-	GatewayTransitionBufferSize int `json:"gateway_transition_buffer_size"`
+	GatewayTransitionBufferSize uint64 `json:"gateway_transition_buffer_size"`
 	// Log is the logger used by the gateway code and gateway packages.
 	Log zerolog.Logger
 	// LogCaller will add the file path and line number to all log messages.
